exynos/fimg2d: increase G2D_MAX_CMD_NR to 110
The kernel currently uses one fourth of the pagesize
for the allocation of the command list backing buffer.

The default pagesize for ARM is 4KB, so the buffer
is 1KB in size. Eight bytes are used for header and
tail of the buffer. Each G2D command consumes eight
bytes, which results in 127 commands.

We currently allow 16 base commands per command list,
hence use 110 as the maximum allowed amount of regular
commands.